requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
weaviate-client==3.24.1
python-dotenv==1.0.0
embed-anything==0.1.1
pydantic==2.4.2
typing-extensions==4.8.0 